+++ bigtree/tree/studio/utils.py
@@ -19,7 +19,7 @@ except ImportError:  # pragma: no cover
     json = MagicMock()
 
 try:
-    from textual.widgets import Static
+    from textual.widgets import Static, Tree
     from textual.widgets._tree import TreeNode
 
 except ImportError:  # pragma: no cover
@@ -50,7 +50,7 @@ SEARCH_EXAMPLES = """[b]Examples[/]
 [cyan]query:age >= 30 OR node_name LIKE ".*e"[/]  Advanced query"""
 
 
-def _get_textual_node_path(textual_node: TreeNode, sep: str = "/") -> str:
+def _get_textual_node_path(textual_node: TreeNode) -> list[str]:
     """Get node path of textual TreeNode, reads the node name recursively until root node.
 
     Args:
@@ -65,7 +65,7 @@ def _get_textual_node_path(textual_node: TreeNode, sep: str = "/") -> str:
     while parent:
         path.append(str(parent.label))
         parent = parent.parent
-    return sep.join(path[::-1])
+    return path[::-1]
 
 
 def _get_corresponding_bt_node(bt_tree: tree.Tree, textual_node: TreeNode) -> node.Node:
@@ -79,7 +79,9 @@ def _get_corresponding_bt_node(bt_tree: tree.Tree, textual_node: TreeNode) -> no
         bigtree tree node
     """
     node_path = _get_textual_node_path(textual_node)
-    return bt_tree.find_full_path(node_path)  # type: ignore[attr-defined, no-any-return]
+    for next_path in node_path[1:]:
+        bt_tree = bt_tree[next_path]
+    return bt_tree.node
 
 
 def _get_corresponding_textual_nodes(
@@ -122,9 +124,21 @@ def _get_attr_bt_node(
     return common.assemble_attributes(bt_node, **kwargs)
 
 
+def _assemble_data(bt_node: node.Node) -> dict[str, str]:
+    """Assemble data for textual node from bigtree node
+
+    Args:
+        bt_node: bigtree node
+
+    Returns:
+        data dictionary
+    """
+    return {"path_name": bt_node.path_name}
+
+
 def populate_textual_tree(
     bt_tree: tree.Tree,
-    textual_tree: TreeNode,
+    textual_tree: Tree,
     max_depth: int = 2,
 ) -> None:
     """Populate textual tree with bigtree tree.
@@ -135,21 +149,18 @@ def populate_textual_tree(
         max_depth: maximum depth of tree to expand
     """
 
-    def assemble_data(bt_node: node.Node) -> dict[str, str]:
-        return {"path_name": bt_node.path_name}
-
     def add(bt_parent: node.Node, textual_parent: TreeNode, depth: int = 1) -> None:
         # Depth + 1 because we are calculating the depth of children
         expand = True if depth + 1 < max_depth else False
         for bt_child in bt_parent.children:
             textual_child = textual_parent.add(
-                bt_child.node_name, data=assemble_data(bt_child), expand=expand
+                bt_child.node_name, data=_assemble_data(bt_child), expand=expand
             )
             add(bt_child, textual_child, depth + 1)
 
     textual_tree.clear()
     textual_tree.root.label = bt_tree.node.name
-    textual_tree.root.data = assemble_data(bt_tree.node)
+    textual_tree.root.data = _assemble_data(bt_tree.node)
     textual_tree.root.expand()
     add(bt_tree.node, textual_tree.root)
 
@@ -191,7 +202,7 @@ def expand_parents(textual_node: TreeNode) -> None:
 
 
 def action_add_node(
-    bt_tree: tree.Tree, textual_node: TreeNode, value: str | None
+    bt_tree: tree.Tree, textual_node: TreeNode | None, value: str | None
 ) -> None:
     """Add new child node to existing node, implements for bigtree tree and textual tree in-place.
 
@@ -203,13 +214,13 @@ def action_add_node(
     if not value or not textual_node:
         return
     bt_node = _get_corresponding_bt_node(bt_tree, textual_node)
-    node.Node(value, parent=bt_node)
-    textual_node.add(value)
+    bt_child = node.Node(value, parent=bt_node)
+    textual_node.add(value, data=_assemble_data(bt_child))
     textual_node.expand()
 
 
 def action_add_sibling(
-    bt_tree: tree.Tree, textual_node: TreeNode, value: str | None
+    bt_tree: tree.Tree, textual_node: TreeNode | None, value: str | None
 ) -> None:
     """Add new sibling node to existing node, implements for bigtree tree and textual tree in-place.
 
@@ -218,16 +229,16 @@ def action_add_sibling(
         textual_node: textual tree node
         value: name of node to add
     """
-    if not textual_node.parent:
-        raise ValueError("Cannot add sibling for root node.")
     if not value or not textual_node:
         return
+    if not textual_node.parent:
+        raise ValueError("Cannot add sibling for root node.")
     bt_node = _get_corresponding_bt_node(bt_tree, textual_node)
-    node.Node(value, parent=bt_node.parent)
-    textual_node.parent.add(value)
+    bt_child = node.Node(value, parent=bt_node.parent)
+    textual_node.parent.add(value, data=_assemble_data(bt_child))
 
 
-def action_delete_node(bt_tree: tree.Tree, textual_node: TreeNode) -> None:
+def action_delete_node(bt_tree: tree.Tree, textual_node: TreeNode | None) -> None:
     """Delete node, implements for bigtree tree and textual tree in-place.
 
     Args:
@@ -256,6 +267,7 @@ def action_rename_node(
     bt_node = _get_corresponding_bt_node(bt_tree, textual_node)
     bt_node.rename(value)
     textual_node.set_label(value)
+    textual_node.data = _assemble_data(bt_node)
 
 
 def select_edit_attr(bt_tree: tree.Tree, textual_node: TreeNode) -> str:
@@ -273,9 +285,9 @@ def select_edit_attr(bt_tree: tree.Tree, textual_node: TreeNode) -> str:
 
 
 def action_edit_attr(
-    bt_tree: tree.Tree, textual_node: TreeNode, value: str | None
+    bt_tree: tree.Tree, textual_node: TreeNode | None, value: str | None
 ) -> None:
-    """Delete node, implements for bigtree tree and textual tree in-place.
+    """Delete node, implements for bigtree tree in-place. Textual tree data only maintains path_name
 
     Args:
         bt_tree: bigtree tree
@@ -284,11 +296,11 @@ def action_edit_attr(
     """
     if not textual_node or not value:
         return
-    kv_pairs = [kv.strip().split("=") for kv in value.split(",")]
-    kv_pairs_eval = [(k, ast.literal_eval(v)) for k, v in kv_pairs]
     try:
+        kv_pairs = [kv.strip().split("=") for kv in value.split(",")]
+        kv_pairs_eval = [(k, ast.literal_eval(v)) for k, v in kv_pairs]
         new_attrs = dict(kv_pairs_eval)
-    except ValueError as err:
+    except (ValueError, SyntaxError) as err:
         raise ValueError(f"Input malformed, check `{value}`") from err
     bt_node = _get_corresponding_bt_node(bt_tree, textual_node)
     existing_attrs = _get_attr_bt_node(bt_tree, textual_node)
@@ -296,6 +308,9 @@ def action_edit_attr(
     for attr_to_remove in attrs_to_remove:
         del bt_node.__dict__[attr_to_remove]
     bt_node.set_attrs(new_attrs)
+    # Sync textual_node to bt_node
+    textual_node.label = bt_node.name
+    textual_node.data = _assemble_data(bt_node)
 
 
 def action_search(
@@ -319,7 +334,7 @@ def action_search(
     return _get_corresponding_textual_nodes(textual_node, bt_matches_path)
 
 
-def action_save_as(bt_tree: tree.Tree, value: str | None) -> None:
+def action_save_as(bt_tree: tree.Tree, value: str | None) -> None:  # pragma: no cover
     """Save tree as json file.
 
     Args: